// src/app/houses/interface.ts
export interface House {
    name: string;
    region: string;
    coatOfArms: string;
    words: string;
    currentLord: string;
    titles: string[];
    seats: string[];
    heir: string;
    overlord: string;
    founded: string;
    founder: string;
    url: string;
    diedOut: string;
    ancestralWeapons: string[];
    cadetBranches: string[];
    swornMembers: string[];
}